requirements.txt
langchain 
openai 
faiss-cpu 
pypdf 
tiktoken 
streamlit
gradio
python-dotenv
pandas
langchain-community
